requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==4.2.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
et_xmlfile==2.0.0
Flask==3.1.0
frozendict==2.4.6
greenlet==3.2.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kaleido==0.2.1
lxml==5.3.2
MarkupSafe==3.0.2
multitasking==0.0.11
numpy==2.2.4
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.7
plotly==5.18.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==13.7
pytimeparse==1.1.8
pytz==2025.2
RapidFuzz==3.13.0
requests==2.32.3
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.40
tenacity==9.1.2
tornado==6.4.2
twelvedata==1.2.11
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
webencodings==0.5.1
Werkzeug==3.1.3